backend/requirements.txt
fastapi~=0.116.1
python-dotenv~=1.0.0
Pillow~=11.3.0
uvicorn==0.35.0
requests
Pillow
numpy
python-multipart
redis
rq
python-dotenv
wandb
opencv-python-headless
tensorflow